Revert "waiting for token"
This reverts commit 3f8ef992d4a3dd0f57eef02df4d45443897c8c1d.
