+++ src/screens/DepositScreen/DepositScreen.tsx
@@ -1,24 +1,16 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Modal, BlankStates } from '@components';
 import { AddFunds } from '@containers';
-import { useDepositProtocols, useTokens } from '@hooks';
+import { useDepositProtocols } from '@hooks';
 import RNUxcam from 'react-native-ux-cam';
-import { MinkeToken } from '@models/token';
 import Deposit from './Deposit/Deposit';
 import OpenSavings from './OpenSavings/OpenSavings';
 import NotAbleToSaveModal from './NotAbleToSaveModal/NotAbleToSaveModal';
 import { useDepositScreen } from './DepositScreen.hooks';
 
 const DepositScreen = () => {
-	const { depositableTokens = [] } = useTokens();
-
-	const {
-		notAbleToSaveVisible,
-		notAbleToSaveDismiss,
-		addFundsVisible,
-		dismissAddFunds, onAddFunds
-	} = useDepositScreen();
-
+	const { notAbleToSaveVisible, notAbleToSaveDismiss, addFundsVisible, dismissAddFunds, onAddFunds } =
+		useDepositScreen();
 	const {
 		setSelectedUSDCoin,
 		depositableToken,
@@ -29,17 +21,6 @@ const DepositScreen = () => {
 		apy
 	} = useDepositProtocols();
 	RNUxcam.tagScreenName('DepositScreen');
-
-	const [token, setToken] = React.useState<MinkeToken>();
-
-	useEffect(() => {
-		if (depositableToken && depositableTokens.length > 0) {
-			const balanceToken = depositableTokens.find(
-				({ symbol }) => symbol.toLowerCase() === depositableToken.symbol.toLowerCase()
-			);
-			setToken(balanceToken);
-		}
-	}, [depositableToken, depositableTokens]);
 
 	if (ableToDeposit === undefined) return <BlankStates.Deposit />;
 
@@ -61,23 +42,8 @@ const DepositScreen = () => {
 	}
 
 	if (approved === undefined) return <BlankStates.Deposit />;
-	if (token === undefined) return <BlankStates.Deposit />;
-	if (approved) {
-		return (
-			<Deposit
-				{...{
-					apy,
-					setSelectedUSDCoin,
-					depositableToken,
-					selectedProtocol,
-					token,
-					setToken,
-					depositableTokens
-				}}
-			/>
-		);
-	}
 
+	if (approved) return <Deposit {...{ apy, setSelectedUSDCoin, depositableToken, selectedProtocol }} />;
 	return <OpenSavings onApprove={() => setApproved(true)} />;
 };
